+++ Frends.HTTP.Request/Frends.HTTP.Request/Definitions/HttpClientFactory.cs
@@ -1,0 +1,13 @@
+﻿using System.Net.Http;
+
+namespace Frends.HTTP.Request.Definitions;
+
+internal class HttpClientFactory : IHttpClientFactory
+{
+    public HttpClient CreateClient(Options options)
+    {
+        var handler = new HttpClientHandler();
+        handler.SetHandlerSettingsBasedOnOptions(options);
+        return new HttpClient(handler);
+    }
+}
